FIX can't install caliper-php lib 2/2
Configured composer.json to download the caliper-php lib archive from
github and install it. Also ran composer update, which changed a bunch
of files. I think those will just get regenerated by composer and don't
have to be checked, but not sure if being a mediawiki extension changes
things, so just going to commit the changes and the new files.