FIX error adding job when wiki in read-only mode
By default, caliper adds jobs to the stock mediawiki job queue. This job
queue is not writeable when the wiki is in read-only mode. So every page
gets an error when they try to emit the view caliper event.

Since this is the stock mediawiki job queue, it's probably not a good
idea to make it writeable in read-only mode. I didn't see any way to
make specific exceptions either in the docs. So the best solution seems
to be to disable caliper when in read-only mode.

If caliper is enabled and we're using the job queue and the wiki is
read-only, caliper will be disable itself.

Note: MW docs recommend using MediaWikiServices::getInstance() directly
only if regular depdendency injection isn't available, such as in static
contexts.

Fixes #3